Scrape action's metadata
Scrape service action's metadata namely, description, access level,
resource types, condition keys and dependent actions.

This is in order to generate extra data on the Actions classes to enable
programatic checking of whether the actions provide the ability to set
condition keys and whether or not the support resource parameters for
limiting access when create and validating least privilege IAM policies.